Adds missing tests for notification type endpoint
